requirements.txt
opencv-python
scikit-learn
pywin32
Pillow